PersonクラスについてREST APIを一通り作成

+++ src/main/java/com/example/demo/repository/PersonRepository.java
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.data.Person;
 
 @Repository
-public interface PersonRepository extends JpaRepository<Person, Long> {}
+public interface PersonRepository extends JpaRepository<Person, Integer> {}
